moved truncate public schema logic into effectstream/db
