+++ packages/node-sdk/runtime/src/main.ts
@@ -7,10 +7,9 @@ import {
   acquireDBMutex,
   createDynamicTables,
   getConnection,
-  getAllTableNames,
   releaseDBMutex,
+  resetPublicTables,
 } from "@effectstream/db";
-import type { IGetAllTableNamesResult } from "@effectstream/db";
 import { PaimaEventBroker } from "@effectstream/event-server";
 import {
   BuiltinEvents,
@@ -188,25 +187,7 @@ function* startup(
 
   // Dev-only reset of user-owned public tables
   if (config.dev?.resetPublicData) {
-    const tableRows = (yield* until(
-      getAllTableNames.run(undefined, dbConn),
-    )) as IGetAllTableNamesResult[];
-    const tableNames = tableRows
-      .map((row) => row.tablename)
-      .filter((name): name is string => Boolean(name));
-    if (tableNames.length > 0) {
-      const qualified = tableNames
-        .map((name) => `"public"."${name}"`)
-        .join(", ");
-      console.log(`[DEV RESET] Truncating public tables: ${qualified}`);
-      yield* until(
-        dbConn.query(
-          `TRUNCATE TABLE ${qualified} RESTART IDENTITY CASCADE`,
-        ),
-      );
-    } else {
-      console.log("[DEV RESET] No public tables found to truncate.");
-    }
+    yield* resetPublicTables(dbConn);
   }
 
   // When the node is started, we apply system migrations.
